+++ commandArgs.js
@@ -26,5 +26,19 @@ module.exports = {
   },
   getArgs() {
     return process.argv.slice(2);
+  },
+  /**
+   * toCommandArgs
+   * @param {Array} - commandArr
+   * @return {Map<String,String>}
+   */
+  toCommandArgs(command = '') {
+    const commandArr = command.split(' ');
+    const map = new Map();
+    commandArr.forEach(t => {
+      const itemArr = t.split('=');
+      map.set(itemArr[0], itemArr[1]);
+    });
+    return map;
   }
 };