+++ commandConfig.js
@@ -3,6 +3,15 @@ const buildapp = require('./buildapp');
 const buildumd = require('./buildumd');
 const buildpackage = require('./buildpackage');
 
+/**
+ * 将用","分割的define参数转换成key/value的map
+ * @param {String} - define
+ * @return Array
+ */
+function getDefineMap(define = '') {
+  return define.split(',');
+}
+
 module.exports = {
   startapp: {
     alias: 'start',
@@ -10,10 +19,17 @@ module.exports = {
     options: [{
       command: '-c, --config <path>',
       description: 'ctbuild.config.js Configuration file path'
+    }, {
+      command: '-d, --define <path>',
+      description: 'custom params split ","'
     }],
     action: (entry) => {
-      console.log('startapp', entry.config);
-      startapp.build(entry.config);
+      console.log('startapp');
+      const {config, define = ''} = entry;
+      startapp.build({
+        config,
+        define: getDefineMap(define)
+      });
     }
   },
   buildapp: {
@@ -22,10 +38,18 @@ module.exports = {
     options: [{
       command: '-c, --config <path>',
       description: 'ctbuild.config.js Configuration file path'
+    }, {
+      command: '-d, --define <path>',
+      description: 'custom params split ","'
     }],
     action: (entry) => {
       console.log('buildapp');
-      buildapp.build(entry.config);
+      // buildapp.build(entry.config, entry.define);
+      const {config, define = ''} = entry;
+      buildapp.build({
+        config,
+        define: getDefineMap(define)
+      });
     }
   },
   buildumd: {
@@ -37,11 +61,18 @@ module.exports = {
     }, {
       command: '-p, --packagename <name>',
       description: 'package name'
+    }, {
+      command: '-d, --define <path>',
+      description: 'custom params split ","'
     }],
     action: (entry) => {
       console.log('buildumd');
-      const {config, packagename} = entry;
-      buildumd.build({config, packagename});
+      const {config, packagename, define = ''} = entry;
+      buildumd.build({
+        config,
+        packagename,
+        define: getDefineMap(define)
+      });
     }
   },
   buildpackage: {