+++ gulpfile.js
@@ -2,7 +2,7 @@ const path = require('path');
 const gulp = require('gulp');
 const uglify = require("gulp-uglify");
 const sourceMap = require('gulp-sourcemaps');
-const copyexts = ['less', 'svg', 'jpg', 'jpeg', 'png', 'bmp', 'json', 'eot', 'woff', 'ttf'];
+const copyexts = ['less', 'css', 'svg', 'jpg', 'jpeg', 'gif', 'png', 'bmp', 'json', 'eot', 'woff', 'ttf'];
 const commandArgs = require('./commandArgs');
 const argsMap = commandArgs.initCommandArgs();
 const outputpath = argsMap.get('--outputpath')[0];
@@ -16,8 +16,8 @@ const compilePath = argsMap.get('--compilepath')[0];
  */
 gulp.task('copy', () => {
   for (let i = 0; i < copyexts.length; i++) {
-    gulp.src(path.join(compilePath,'**',`*.${copyexts[i]}`))
-        .pipe(gulp.dest(outputpath));
+    gulp.src(path.join(compilePath, '**', `*.${copyexts[i]}`))
+      .pipe(gulp.dest(outputpath));
   }
 });
 
@@ -28,8 +28,8 @@ gulp.task('minjs', () => {
   return gulp.src([
     // `${runtimePath}lib\\**\\*.js`,
     // `${runtimePath}lib\\**\\*.jsx`,
-    path.join(outputpath,'**','*.js'),
-    path.join(outputpath,'**','*.jsx'),
+    path.join(outputpath, '**', '*.js'),
+    path.join(outputpath, '**', '*.jsx'),
   ]).pipe(sourceMap.init())
     .pipe(uglify())
     .pipe(sourceMap.write('.'))