+++ startapp.js
@@ -9,6 +9,7 @@ const codePath = __dirname;
 const commandPath = path.join(codePath, 'node_modules', '.bin', path.sep);
 // 配置文件所在路径
 let configPath;
+let define;
 
 /**
  * corssenvTask
@@ -92,7 +93,9 @@ function webpackServiceTask() {
         '--runtimepath',
         path.join(runtimePath, path.sep),
         '--customconfig',
-        configPath
+        configPath,
+        '--define',
+        define.join(' ')
       ],
       {
         cwd: codePath,
@@ -150,7 +153,7 @@ module.exports = {
    * @param {String} - ctbuildConfigPath
    * ctbuild.config.js配置文件的路径，如果没有指定则会寻找命令运行目录下的ctbuild.config.js文件
    */
-  build: (ctbuildConfigPath = '') => {
+  build: ({config: ctbuildConfigPath = '', define: defineMap}) => {
     if(ctbuildConfigPath) {
       if(path.isAbsolute(ctbuildConfigPath)) {
         configPath = ctbuildConfigPath;
@@ -160,6 +163,8 @@ module.exports = {
     } else {
       configPath = path.join(runtimePath,'ctbuild.config.js');
     }
+
+    define = defineMap;
 
     loopTask().then(() => {
       console.log('finish');